refactored imports, upgraded vite
