+++ PlantDB-Frontend/package.json
@@ -41,6 +41,6 @@
     "eslint-plugin-react-refresh": "^0.3.4",
     "icons": "link:@types/@tabler/icons",
     "typescript": "^5.0.2",
-    "vite": "^4.3.9"
+    "vite": "^4.4.9"
   }
 }
