+++ PlantDB-Frontend/src/components/InputGroup.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from "react";
-import { css, cx } from "../../styled-system/css";
+import { css, cx } from "styled-system/css";
 
 type InputGroupProps = {
   /**
