+++ PlantDB-Frontend/src/sections/entries/add/index.tsx
@@ -1,19 +1,18 @@
 import { useNavigate, useParams } from "react-router-dom";
-import { css } from "../../../../styled-system/css";
 import { useTranslation } from "react-i18next";
 import { useMutation, useQuery, useQueryClient } from "react-query";
 import { useForm, type SubmitHandler } from "react-hook-form";
-import { type PlantRecord, type PlantInfo } from "../../../types";
-import { useToast } from "../../../components/Toast";
-import EntryForm from "../EntryForm";
-import Loading from "../../../components/Loading";
-import HeaderBar from "../../../components/HeaderBar";
-import Breadcrumbs from "../../../components/Breadcrumbs";
-import Container from "../../../components/Container";
-import FormSkeleton from "../FormSkeleton";
-import getRequestFormData from "../../../helpers/getRequestFormData";
-import Logo from "../../../components/Logo";
 import { Helmet } from "react-helmet";
+import { css } from "styled-system/css";
+import { type PlantRecord, type PlantInfo } from "@/types";
+import { useToast } from "@/components/Toast";
+import Loading from "@/components/Loading";
+import HeaderBar from "@/components/HeaderBar";
+import Breadcrumbs from "@/components/Breadcrumbs";
+import Container from "@/components/Container";
+import getRequestFormData from "@/helpers/getRequestFormData";
+import Logo from "@/components/Logo";
+import EntryForm from "../EntryForm";
 
 function Add() {
   const queryClient = useQueryClient();
@@ -100,27 +99,3 @@ function Add() {
 }
 
 export default Add;
-
-export function Skeleton() {
-  const { species } = useParams();
-  const { t } = useTranslation("entries");
-
-  return (<>
-    <HeaderBar>
-      <Logo />
-      <Breadcrumbs links={[
-        {
-          title: t(species?.toLocaleLowerCase() + ".plural"),
-          to: `/${species}`
-        },
-        {
-          title: t("add.title")
-        }
-      ]} />
-    </HeaderBar>
-
-    <Container>
-      <FormSkeleton />
-    </Container>
-  </>);
-}
