+++ PlantDB-Frontend/tsconfig.json
@@ -18,7 +18,12 @@
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true
+    "noFallthroughCasesInSwitch": true,
+
+    "paths": {
+      "@/*": ["./src/*"],
+      "styled-system/*": ["./styled-system/*"]
+    }
   },
   "include": ["src", "styled-system"],
   "references": [{ "path": "./tsconfig.node.json" }]
